backend/requirements/base.txt
# Core
Django==5.2.4
tzdata==2025.2
psycopg[binary]<4.0,>=3.1

# API
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2.2
drf-spectacular>=0.27.1
PyYAML>=6.0

# Static files
whitenoise>=6.6.0

# Utils
python-dotenv>=1.0
